Add support for k-means clustering dominant colors algorithm.
This commit offers another way of calculating the dominant colors of an image: k-means clustering.
The results are a bit more iffy in general, but it offers the advantage of having much more content computational times than the iterative approach.
It also has the advantage of being backed by a first party API (the `CIKMeans` filter).